third_party/blink/web_tests/external/wpt/html/semantics/embedded-content/the-iframe-element/iframe-loading-lazy-in-script-disabled-iframe-expected.txt

This is a testharness.js-based test.
FAIL Iframes with loading='lazy' in script disabled iframe are not handled
       as 'lazy' assert_equals: lazy-load iframe shouldn't be honored in script disabled iframe expected "Subframe" but got ""
Harness: the test ran to completion.
